refactor: minor improvements

+++ test/BaseTest.sol
@@ -57,8 +57,8 @@ contract BaseTest is Test {
         vault.setVic(address(vic));
     }
 
-    function bound96(uint96 x, uint96 _min, uint96 _max) public pure returns (uint96) {
-        return uint96(bound(uint256(x), uint256(_min), uint256(_max)));
+    function bound96(uint96 x, uint256 _min, uint256 _max) public pure returns (uint96) {
+        return uint96(bound(uint256(x), _min, _max));
     }
 }
